View character ranges in a specific font
Shows only the codepoints that are defined in a specific font (and
marked as normal, printable characters in Unicode).

For example:

	fonttable -f MyriadPro

Note that your machine must have FontConfig and the fontname must in
a format FontConfig likes. For example:

	fonttable -f LTCCaslon				# Works
	fonttable -f "LTC Caslon"			# Works
	fonttable -f "LTC Caslon Swash Long Regular"	# Works

But:
	fonttable -f Caslon			# Does Not Work
	fonttable -f "LTC Caslon Swash Long"	# Does Not Work

____
Also note that some decorative and expert fonts define glyphs at
unusual character locations which are not standard Unicode. Those
codepoints will not be found by this script which only prints
Unicode. (But, that could change in the future if there is a need.)

For example:

	fonttable -f AdobeWoodType		# Shows no glyphs!